% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBasePAR.r
\name{createBasePAR}
\alias{createBasePAR}
\title{createBasePAR}
\usage{
createBasePAR(parFile, base = "../model/formind_parameters/", pattern = NA,
  replacement = NA)
}
\arguments{
\item{parFile}{Character sting naming the new par-file.}

\item{base}{Character sting naming the already existing par-file.}

\item{pattern}{Vector of character strings naming those variables of the par file which will be modified}

\item{replacement}{Vector of character strings which will replace the line, where the pattern is found.}
}
\description{
create a modifyable par file based on another parfile.
}
