\name{FAST}
\alias{FAST-package}
\alias{FAST}
\docType{package}
\title{
FORMIND analysis and simulation tool
}
\description{
This package was design to analyse and simulate virtual forest with the gap model \href{http://formind.org}{FORMIND}. Nevertheless, it can be used with data of other models or even data of real forests. 
}
\details{
\tabular{ll}{
Package: \tab FAST\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-06-08\cr
License: \tab GPL-3\cr
}
FAST introduces new class: Forest. FAST includes several functions for generation of Forests as well as plotting information of Forests
}
\author{
Friedrich J. Bohn
 friedrich.bohn@ufz.de
}
\references{
in prep Bohn et al. 2016
}
\keyword{ package }
%%\seealso{
%%}
%%\examples{

%%}
