% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.r
\name{fast}
\alias{fast}
\title{fast}
\usage{
fast(model, param, runtime = 500, repetitions = 1, suc.plot = T,
  circle.plot = F, hist.plot = F, graficType = "paperC2",
  newSimulation = T)
}
\arguments{
\item{model}{Character sting naming the model-file.}

\item{param}{Character sting naming the par-file.}

\item{runtime}{Number of years to simulate.}

\item{newSimulation}{Boolean; Default TRUE; if FALSE no simulation is processed but data are analysed if available.}
}
\description{
conducting a simulation and analyse it in one go
}
