% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastLmX.R
\name{fastLmX}
\alias{fastLmX}
\alias{fastLmX.default}
\alias{print.fastLmX}
\alias{summary.fastLmX}
\alias{fastLmX.formula}
\alias{predict.fastLmX}
\title{fastLmX: Weighted Regression Rewritten using RCpp}
\usage{
fastLmX(X, ...)

\method{fastLmX}{default}(X, Y, Z = replicate(length(y), 1))

\method{print}{fastLmX}(x, ...)

\method{summary}{fastLmX}(object, ...)

\method{fastLmX}{formula}(formula, data = list(), ...)

\method{predict}{fastLmX}(object, newdata = NULL, ...)
}
\arguments{
\item{X}{a data.frame consists of explantory variables}

\item{Y}{a Vector consists of output variables}

\item{Z}{a Vector consists of weights}
}
\description{
A description of fastLmX
}
\details{
A details of fastLmX
}
\examples{
class(a) <- "data.frame"
class(b) <- "numeric"
class(c) <- "numeric"
fastLmX(a, b, c)

}
