% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastLmX_impl}
\alias{fastLmX_impl}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
fastLmX_impl(X, Y, Z)
}
\arguments{
\item{Y}{a Numeric Vector}

\item{Z}{a Numeric Vector for weight}

\item{x}{a Matrix}
}
\description{
estimate the regression estimates based on given the number of repetitions
}
