% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forest_Class.r
\docType{class}
\name{Forest-class}
\alias{Forest-class}
\title{An S4 class to represent a bank account.}
\description{
An S4 class to represent a bank account.
}
\section{Slots}{

\describe{
\item{\code{balance}}{A length-one numeric vector}
}}

