% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getForestColNames.R
\name{getForestColNames}
\alias{getForestColNames}
\title{getForestColNames}
\usage{
getForestColNames(forest)
}
\arguments{
\item{forest}{Forest object;}
}
\description{
returns the colnames of the data within the forest object
}
