% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectPath.r
\name{getProjectPath}
\alias{getProjectPath}
\title{getProjectPath}
\usage{
getProjectPath(FolderName)
}
\arguments{
\item{FolderNAme}{string, naming the project folder.}
}
\description{
function only needed for FORMIND-tutorial
}
