% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{lm1_X}
\alias{lm1_X}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
lm1_X(formula, data, freqs)
}
\description{
estimate the regression estimates based on given the number of repetitions
}
