% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeColorPalette.R
\name{makeColorPalette}
\alias{makeColorPalette}
\title{create various colour palettes}
\usage{
makeColorPalette(n, graphicType = "plotBox", totalExists = F)
}
\arguments{
\item{n}{Number of colours}

\item{graphicType}{For which grafic type should the colours be optimized ("paperC1","paperC2","plotBox","ppt")}

\item{totalExists}{Boolean. If TRUE black is added to the pallete.}
}
\description{
This function creates various colour pallets which suit to the different layouts
}
