% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLayout.R
\name{makeLayout}
\alias{makeLayout}
\title{makeLayout}
\usage{
makeLayout(graficType, innerTicks = T, las = 1, file = NA)
}
\arguments{
\item{graficType}{Character string; "paperC1" optimizes graphic for one collumen in an a paper. "paperC1" optimizes for a graphic straging over two collumns in a paper. "ppt" optimizes for powerpoint. "plotWindow" creates a new grahic device (window).}

\item{innerTicks}{boolean; see par}

\item{las}{numeric; see par}

\item{file}{character string; filename where graphic will be saved. allowed suffixes:"png", "pdf".}
}
\description{
male a nice layout
}
