% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLegend.R
\name{makeLegend}
\alias{makeLegend}
\title{intern: add a shiny legend}
\usage{
makeLegend(labels, farben, graficType, varNames, y)
}
\arguments{
\item{labels}{describtion of labels}

\item{farben}{the collors}

\item{graficType}{the grafic type}

\item{varNames}{the names}

\item{y}{the y}
}
\description{
modify and rename par-files
}
