% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pftRainbow.R
\name{pftRainbow}
\alias{pftRainbow}
\title{pftRainbow}
\usage{
pftRainbow(n, alpha = NA, lum.fix = FALSE)
}
\arguments{
\item{n}{numeric; number of PFTs (colours)}

\item{alpha}{numeric; between 0 & 1}

\item{lum.fix}{bool; fix the luminicenc variable}
}
\description{
create nice rainbow palette.
}
