% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSuccession.R
\name{plotSuccession}
\alias{plotSuccession}
\title{plot formind simulation data over time}
\usage{
plotSuccession(data, y, col = NA, graphicType = "paperC2",
  line_shows = "mean", band = NA, file = NA, firstLetterCapital = F,
  squaredBraked = T, innerTicks = T, las = 0, legend = T, xpos = 0,
  ypos = 0)
}
\arguments{
\item{data}{A forest object including the data}

\item{y}{Vector of character strings nameing the variables which will be plotted. \code{getFNames(object)} returns an overview of available variables. You can also write short strings to plot all variables which start with these characters. E.g. "biomass_pft" plots the biomass of all pfts, whereas "biomass" plots biomass of all pfts and the total biomass.}

\item{col}{Color palette to be used. If NA, an adequate collorpallet is created}

\item{graphicType}{Character string indicating layout type of the graphic. "ppt" optimizes for powerpoint slides; "paperC1","paperC2" optimizes for graphic stretching over one or two collumns within a paper; "plotbox" opens a window an optimizes for the screen.}

\item{firstLetterCapital}{Boolean. Should first Letter of labales be capitalized}

\item{squaredBraked}{Boolean. squared brakets (TRUE) or round brakets (FALSE) for unit information in the labels}

\item{innerTicks}{Boolean.}

\item{las}{numeric in {0,1,2,3}; the style of axis labels. 0: always parallel to the axis [default],1: always horizontal,2: always perpendicular to the axis, 3: always vertical.}

\item{lineShows}{Character sting indicate what is shown by the line: "mean" or "median"}

\item{legend.}{Boolean. Should the legend be plotted.}

\item{timeframe}{A vector of the form \code{c(startTime,endTime)}.}
}
\description{
Process forest-class-objects to create time vs. variable plots.
}
\seealso{
\code{\link{Forest}},
}
