% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCMD.r
\name{prepareCMD}
\alias{prepareCMD}
\title{prepareCMD}
\usage{
prepareCMD(file, formindVersion = "formind.exe", parFile = "experiment.par")
}
\arguments{
\item{file}{Character string naming the cmd.file}

\item{formindVersion}{Character string indicating the FORMIND-executable relative to the location of cmd.file, which will be used}

\item{parFile}{Character string indicating the par-file in the formind_parameters folder which must be at the same location as the cmd-file.}
}
\description{
writes the CMDfile, which allows to run a formind with a given par-file.
}
