% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFORMIND.R
\name{runFORMIND}
\alias{runFORMIND}
\title{runFORMIND}
\usage{
runFORMIND(model = "formind.exe", parFile = "experiment.par",
  parStandardPath = T)
}
\arguments{
\item{model}{Character string naming the modelfile}

\item{parFile}{Character string naming the PARfile}

\item{parStandardPath}{bool; if T parfile is in folder formind_parameters}
}
\description{
runs FORMIND using a CMDfile and check run for errors.
}
\seealso{
\code{\link{prepareCMD}}
}
