% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writingPINfile.R
\name{writePINfile}
\alias{writePINfile}
\title{writingPINfile}
\usage{
writePINfile(data, dbhclasses, filename, seedData = NA, no_pft = NA)
}
\arguments{
\item{data}{a matrix: plotnumber,species (as number),which dbhclass as index}

\item{dbhclasses}{vector containing the dbh classes}

\item{filename}{string of the new pin-file}

\item{seedData}{string of the seed in every plot}

\item{no_pft}{number of pfts.}
}
\description{
writes a pin file
}
