% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\docType{package}
\name{fast-package}
\alias{fast}
\alias{fast-package}
\title{fast: fast computational weighted linear Regression}
\description{
Fast Computational weighted linear Regression
    Using Linear Algebra
}
\details{
Linear Regression with Little Bag of Bootstraps
}
