% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histStemSizeDist.R
\name{histStemSizeDist}
\alias{histStemSizeDist}
\title{make a histogram}
\usage{
histStemSizeDist(data, y = "NumberTreesTotal", col = NA, breaks = 0,
  usedYears = 0.1, graphicType = "plotBox", xlim = NA, ylim = NA,
  innerTicks = T, file = NA, las = 1, log = NA, legend = T)
}
\arguments{
\item{data}{A forest object including the data}

\item{y}{Vector of character strings nameing the variables which will be plotted. \code{getFNames(object)} returns an overview of available variables. You can also write short strings to plot all variables which start with these characters. E.g. "biomass_pft" plots the biomass of all pfts, whereas "biomass" plots biomass of all pfts and the total biomass.}

\item{col}{Color palette to be used. If NA, an adequate collorpallet is created}

\item{breaks}{one of:
(i)a vector giving the breakpoints between histogram cells,
(ii) a function to compute the vector of breakpoints,
(iii) a single number giving the number of cells for the histogram,
(iv)a character string naming an algorithm to compute the number of cells (see ‘Details’),
(v) a function to compute the number of cells.
In the last three cases the number is a suggestion only; the breakpoints will be set to pretty values. If breaks is a function, the x vector is supplied to it as the only argument.}

\item{usedYears}{numeric; the last x persantage [0-1] of the time line used for analysis}

\item{innerTicks}{Boolean.}

\item{las}{numeric in {0,1,2,3}; the style of axis labels. 0: always parallel to the axis [default],1: always horizontal,2: always perpendicular to the axis, 3: always vertical.}

\item{log}{string specifying if axis scales should be logarithmic}

\item{legend}{Boolean. Should the legend be plotted.}

\item{firstLetterCapital}{Boolean. Should first Letter of labales be capitalized}

\item{squaredBraked}{Boolean. squared brakets (TRUE) or round brakets (FALSE) for unit information in the labels}

\item{legend.}{Boolean. Should the legend be plotted.}

\item{timeframe}{A vector of the form \code{c(startTime,endTime)}.}

\item{timeframe}{A vector of the form \code{c(startTime,endTime)}.}
}
\description{
histogramms of the data of an forest object
}
\seealso{
\code{\link{Forest}},
}
