% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPAR.R
\name{readPAR}
\alias{readPAR}
\title{read  par-files}
\usage{
readPAR(parFile, parameter = NA, pftNames = NA)
}
\arguments{
\item{parFile}{Character string indicating the par-file, which will be read}

\item{parameter}{vector of parameter files or NA. default: NA ,reads all parameters}
}
\description{
read par-files to R
}
\note{

}
